Restructure the interface to JointLimit template
This adds class template arguments to pass the BM statements into the
class template. It is necessary to do it this way, because define
statements do not survive into the class templates otherwise.
This all removes the Enable_All_JointLimits argument. Now the
joint limits can be disabled individually instead.